Fixing build dependencies

+++ kernel/arch/x86_64/paging_test_utils.c
@@ -1,4 +1,4 @@
-#include <kernel/arch/x86_64/paging_test_utils.h>
+#include <arch/x86_64/paging_test_utils.h>
 
 #include <kernel.h>
 #include <string.h>
@@ -36,7 +36,7 @@ void write_pte(lpaddr_t source, lpaddr_t dest, bool valid){
 void read_memory(lvaddr_t addr){
 }
 
-void write_memory(lvaddr_t addr, uint64 value){
+void write_memory(lvaddr_t addr, uint32_t value){
 }
 
 void invalidate_page(lvaddr_t addr){
